requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-heroku==0.3.1
django-livereload-server==0.4
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
nexmo==2.5.2
oauthlib==3.2.1
phonenumbers==8.12.56
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
twilio==7.15.0
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
